Moves from RuntimeStringInterpolator to MacroInterpolator.
This change allows throwing discrepancy errors between queries and model at compile time instead of
runtime.

It additionally splits the application into two different SBT modules, one for the examples and one
for the models.

We are doing it to avoid issues with macro expansion, which is additionally defining a new macro
(specifically the contextual macro interpolator), so we need them in different compilation units to
make it visible to the application code.